+++ {{cookiecutter.directory_name}}/apps/api/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from apps.api.views import status, user, token, password
+from apps.api.views import status, user, token, recovery_code
 
 urlpatterns = [
     # Authentication
@@ -11,9 +11,9 @@ urlpatterns = [
     path('users/<uuid:user_id>', user.UserDetail.as_view()),
     path('users/me', user.UserMe.as_view()),
 
-    # Password Recovery
-    path('password/recovery', password.PasswordRecoveryManagement.as_view(), name='password-recovery'),
-    path('password/change', password.PasswordChangeManagement.as_view(), name='password-change'),
+    # Recovery Code
+    path('recovery_code', recovery_code.RecoveryCodeManagement.as_view(), name='recovery-code'),
+    path('recovery_code/<uuid:recovery_code_id>', recovery_code.RecoveryCodeDetail.as_view(), name='recovery-code-id'),
 
     # Status
     path("status", status.StatusManagement.as_view(), name='status'),