+++ {{cookiecutter.directory_name}}/apps/api/views/recovery_code.py
@@ -2,6 +2,7 @@ import random
 import time
 from datetime import timedelta
 from http import HTTPStatus
+from uuid import UUID
 
 from django.conf import settings
 from django.db import transaction
@@ -10,22 +11,22 @@ from django.utils import timezone
 from django.utils.translation import gettext as _
 
 from apps.api.errors import ValidationException, ProblemDetailException
-from apps.api.forms.password import RecoveryCodeForm, NewPasswordForm
+from apps.api.forms.recovery_code import RecoveryCodeForm
 from apps.api.response import SingleResponse
 from apps.api.views.base import SecuredView
 from apps.core.models import User, RecoveryCode
 from apps.core.services.email_notification import NotificationEmailService
 
 
-class PasswordRecoveryManagement(SecuredView):
+class RecoveryCodeManagement(SecuredView):
     EXEMPT_AUTH = ['POST']
 
     @transaction.atomic
     def post(self, request):
-        form = RecoveryCodeForm.create_from_request(request)
+        form = RecoveryCodeForm.Email.create_from_request(request)
 
         if not form.is_valid():
-            raise ValidationException(request, form)
+            raise ValidationException(form)
 
         try:
             user = User.objects.get(email=form.cleaned_data['email'])
@@ -42,7 +43,9 @@ class PasswordRecoveryManagement(SecuredView):
             subject=_('Change password'),
             content={
                 'recovery_code': recovery_code,
-                'login_url': f'{settings.BASE_URL}{reverse("password-change")}',
+                'login_url': request.build_absolute_uri(
+                    reverse('recovery-code-id', kwargs={'recovery_code_id': recovery_code.pk})
+                ),
             },
             template=settings.EMAIL_RECOVERY_PATH,
         ).send_email()
@@ -50,25 +53,26 @@ class PasswordRecoveryManagement(SecuredView):
         return SingleResponse(request)
 
 
-class PasswordChangeManagement(SecuredView):
+class RecoveryCodeDetail(SecuredView):
     EXEMPT_AUTH = ['POST']
 
     @transaction.atomic()
-    def post(self, request):
-        form = NewPasswordForm.create_from_request(request)
+    def post(self, request, recovery_code_id: UUID):
+        form = RecoveryCodeForm.Password.create_from_request(request)
 
         if not form.is_valid():
-            raise ValidationException(request, form)
+            raise ValidationException(form)
 
         RecoveryCode.objects.filter(
             created_at__lte=timezone.now() - timedelta(days=settings.RECOVERY_TIME_EXPIRATION)
         ).delete()
 
         try:
-            recovery_code = RecoveryCode.objects.get(pk=form.cleaned_data['recovery_code'])
+            recovery_code = RecoveryCode.objects.get(pk=recovery_code_id)
         except RecoveryCode.DoesNotExist:
             raise ProblemDetailException(
-                request, _('Recovery code expired or does not exists.'), status=HTTPStatus.NOT_FOUND
+                _('Recovery code expired or does not exists.'),
+                status=HTTPStatus.NOT_FOUND
             )
 
         recovery_code.user.set_password(form.cleaned_data['password'])