+++ {{cookiecutter.directory_name}}/{{cookiecutter.project_name}}/settings/production.py
@@ -5,3 +5,5 @@ TIME_ZONE = 'Europe/Bratislava'
 ALLOWED_HOSTS = [
     ''
 ]
+
+SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")